Japanese Pages - About Us
* Fixed header & footer error
* Made Fragment Modal Click handler generic
