+++ blocks/cards/cards.js
@@ -1,5 +1,6 @@
 import { createOptimizedPicture } from '../../scripts/lib-franklin.js';
 import { cropString } from '../../scripts/scripts.js';
+import {handleModalClick} from '../../scripts/scripts.js';
 
 export default function decorate(block) {
   /* change to ul, li */
@@ -44,6 +45,11 @@ export default function decorate(block) {
       [...a.children].forEach(addCardChildrenClasses);
     } else {
       [...li.children].forEach(addCardChildrenClasses);
+    }
+
+    if (a && block.classList.contains('video-modal') && document.querySelector('.modal-fragment') !== null) {
+      // Videos embedded through Fragments Modals will be opened in a modal
+      handleModalClick(a, document.querySelector('.modal-fragment'));
     }
 
     const title = li.querySelector('.title');
