+++ blocks/columns/columns.css
@@ -36,6 +36,12 @@
   width: auto;
 }
 
+.columns.left-section-half-width > div > div > div {
+  width: 50%;
+  margin-left: auto;
+  margin-right: auto;
+}
+
 .columns.small-margin > div {
   margin-bottom: 1.25rem;
 }
@@ -226,11 +232,11 @@
   gap: 0;
 }
 
-.columns.text-center > div div.text-col p, 
-.columns.text-center > div div.text-col h2, 
-.columns.text-center > div div.text-col h3, 
-.columns.text-center > div div.text-col h4, 
-.columns.text-center > div div.text-col h5, 
+.columns.text-center > div div.text-col p,
+.columns.text-center > div div.text-col h2,
+.columns.text-center > div div.text-col h3,
+.columns.text-center > div div.text-col h4,
+.columns.text-center > div div.text-col h5,
 .columns.text-center > div div.text-col h6 {
   text-align: center;
 }
